refactor(lib): changes to the code to publish a compiled version to npm
