+++ .babelrc
@@ -1,0 +1,6 @@
+{
+  "presets": [
+    ["@babel/preset-env", { "useBuiltIns": "entry" }],
+    ["react-app", { "absoluteRuntime": false }]
+  ]
+}
