+++ package.json
@@ -3,19 +3,22 @@
   "version": "0.0.0-development",
   "private": false,
   "license": "BSD-3-Clause",
+  "main": "dist/index.js",
+  "module": "dist/index.js",
+  "files": [
+    "dist",
+    "README.md"
+  ],
   "dependencies": {
     "@material-ui/core": "3.9.2",
     "@material-ui/icons": "3.0.2",
     "lodash": "4.17.11",
     "moment": "2.24.0",
-    "react": "16.8.5",
-    "react-dom": "16.8.5",
     "react-popper": "1.3.3",
-    "react-scripts": "2.1.8",
     "styled-components": "4.2.0"
   },
   "scripts": {
-    "build": "react-scripts build",
+    "build": "rm -rf dist && NODE_ENV=production babel src/lib --out-dir dist --copy-files --ignore __tests__,spec.js,test.js,__snapshots__",
     "commit": "npx git-cz",
     "eject": "react-scripts eject",
     "format:all": "prettier .prettierrc --write '**/*.*(js|jsx|md|mdx)'",
@@ -42,12 +45,17 @@
   "resolutions": {
     "ansi-styles": "^3.2.0"
   },
+  "peerDependencies": {
+    "react": "16.8.5",
+    "react-dom": "16.8.5"
+  },
   "devDependencies": {
+    "@babel/cli": "~7.4.3",
     "@emotion/core": "10.0.6",
     "docz": "0.13.7",
     "docz-theme-default": "0.13.7",
-    "emotion-theming": "10.0.6",
     "dom-testing-library": "3.19.1",
+    "emotion-theming": "10.0.6",
     "eslint-plugin-flowtype": "3.6.1",
     "eslint-plugin-react": "7.12.4",
     "faker": "4.1.0",
@@ -57,6 +65,7 @@
     "lint-staged": "8.1.5",
     "prettier": "1.17.0",
     "pretty-quick": "1.10.0",
+    "react-scripts": "2.1.8",
     "react-testing-library": "6.1.2",
     "semantic-release": "15.13.3",
     "user-event": "1.4.8"
