+++ src/lib/common/__tests__/DiscourseService.spec.js
