+++ src/lib/common/helpers.js
