+++ src/lib/locales/index.js
