+++ src/lib/notifications/__tests__/normaliser.spec.js
