+++ src/lib/profile/__tests__/UserAvatarDropdown.spec.js
