+++ src/lib/session/__tests__/Session.spec.js
