+++ src/lib/session/__tests__/SessionService.spec.js
